initiatives/agent_security_initiative/frameworks/autogen/autogen_refund_demo/requirements.txt
autogen-agentchat==0.2.40
openai>=1.3,<1.58.0
llama-index>=0.12.8
python-dotenv>=0.21.1 
pydantic>=2.0.0,<3.0.0